Moved benchmark to test source
